+++ HISTORY.rst
@@ -4,7 +4,8 @@ History
 2.2.10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Add `substances` and overwrite get_model_instance_by_field_name method
+  in GridH5WaterQualityResultAdmin.
 
 
 2.2.9 (2024-05-27)